limit provider connections number

+++ provider/src/main/java/com/example/provider/CoronaSummaryDto.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class CoronaSummaryDto {
-    private Long confirmedCases;
-    private Long deaths;
+    private Integer confirmedCases;
+    private Integer deaths;
 }
